feat: 문제 답변 페이지 헤더 컴포넌트 분리
