+++ src/app/questions/[questionid]/page.tsx
@@ -1,12 +1,10 @@
 "use client";
 import { useRef, useState } from "react";
-import Image from "next/image";
-import { Badge } from "@/components/ui/badge";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Button } from "@/components/ui/button";
-
 import { useAuthStore } from "@/store/useAuthStore";
 import AnswerPreviewCard from "@/app/api/answers/componsts/AnswerPreviewCard";
+import QusetionHeader from "@/app/api/answers/componsts/QusetionHeader";
 
 interface Props {
   params: {
@@ -19,13 +17,13 @@ export default function AnswerFormPage({ params }: Props) {
   const questionId = Number(params?.questionid);
   const answerRef = useRef<HTMLTextAreaElement>(null);
   const [tab, setTab] = useState<string>("tab1");
-  const [isBookmarked, setIsBookmarked] = useState<boolean>(false);
+
   const [isSubmit, setIsSubmit] = useState(false);
   const [isEditing, setIsEditing] = useState(true);
   const token = useAuthStore((state) => state.token);
   const user = useAuthStore((state) => state.user);
   const userEmail = user?.email;
-  const handleToggleBookmark = () => setIsBookmarked((prev) => !prev);
+
   //이전에 작성한 답변 불러오기
 
   //답변 저장, 수정
@@ -38,7 +36,7 @@ export default function AnswerFormPage({ params }: Props) {
     }
     const formData = {
       userId: userEmail,
-      questionId: 6,
+      questionId,
       content,
     };
     try {
@@ -89,25 +87,7 @@ export default function AnswerFormPage({ params }: Props) {
 
   return (
     <div className="w-[1270px] container mx-auto pt-[40px]">
-      <header className="flex justify-between items-center pb-[18px]">
-        <div className="flex items-center pb-[18px]">
-          <Badge className="mr-[16px]">Javascript</Badge>
-          <h3 className="txt-3xl-b">HTTP 메소드에 대한 설명</h3>
-        </div>
-        <div
-          className="flex items-center justify-center w-[32px] h-[32px]"
-          onClick={handleToggleBookmark}
-        >
-          <Image
-            src={`/assets/icons/bookmark${isBookmarked ? "_fill" : ""}.svg`}
-            alt="bookmark icon"
-            width={24}
-            height={24}
-            className={`w-[24px] h-[24px] object-center cursor-pointer ${isBookmarked && "w-[28px] h-[28px]"}`}
-          />
-        </div>
-      </header>
-
+      <QusetionHeader />
       <form>
         <Tabs defaultValue="tab1" value={tab} onValueChange={setTab}>
           <TabsList className="mr-0 ml-auto">
